use .scipy file to tell buildstep to use scipy config

+++ upload.sh
@@ -1,8 +1,16 @@
 #/bin/bash
+echo -n "Are you sure you want to upload a new buildstep image? [yn] "
+read -n 1 confirm
+[[ "$confirm" = "y" ]] || exit 0
+
 [[ -f /tmp/tgz ]] && rm /tmp/tgz
 ID=$(docker run -d progrium/buildstep /bin/sh)
 SHORTID=${ID:0:10}
-DESTINATION=s3://expa-dokku/expa_buildstep_${SHORTID}.tgz
+if [ -e './stack/.scipy' ];then
+  DESTINATION=s3://expa-dokku/expa_scipy_buildstep_${SHORTID}.tgz
+else
+  DESTINATION=s3://expa-dokku/expa_buildstep_${SHORTID}.tgz
+fi
 
 echo "exporting $ID"
 docker export $ID | gzip -9c > /tmp/tgz || exit 1
